+++ api/lock_api_test.go
@@ -1,9 +1,7 @@
 package api_test
 
 import (
-	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"testing"
 	"time"
 
@@ -14,6 +12,8 @@ import (
 var LockService api.LockService
 
 func TestSuccessfullyObtainingALock(t *testing.T) {
+	now := time.Now()
+
 	schema, resp := LockService.Lock(&api.LockRequest{
 		Path:               "/path/to/file",
 		LatestRemoteCommit: "deadbeef",
@@ -25,10 +25,10 @@ func TestSuccessfullyObtainingALock(t *testing.T) {
 
 	tc := &MethodTestCase{
 		Schema:         schema,
-		Response:       resp,
 		ExpectedPath:   "/locks",
 		ExpectedMethod: http.MethodPost,
-		ExpectedResponse: &api.LockResponse{
+		Resp:           resp,
+		Output: &api.LockResponse{
 			Lock: api.Lock{
 				Id:   "some-lock-id",
 				Path: "/path/to/file",
@@ -37,55 +37,18 @@ func TestSuccessfullyObtainingALock(t *testing.T) {
 					Email: "jane@example.com",
 				},
 				CommitSHA: "deadbeef",
-				LockedAt:  time.Date(2016, time.May, 18, 0, 0, 0, 0, time.UTC),
+				LockedAt:  now,
 			},
 		},
-		Output: `
-{
-	"lock": {
-		"id": "some-lock-id",
-		"path": "/path/to/file",
-		"committer": {
-			"name": "Jane Doe",
-			"email": "jane@example.com"
-		},
-		"commit_sha": "deadbeef",
-		"locked_at": "2016-05-18T00:00:00Z"
-	}
-}`,
 	}
 
 	tc.Assert(t)
-}
 
-type MethodTestCase struct {
-	Schema   *api.RequestSchema
-	Response interface{}
-
-	ExpectedPath     string
-	ExpectedMethod   string
-	ExpectedResponse interface{}
-
-	Output string
-}
-
-func (c *MethodTestCase) Assert(t *testing.T) {
-	var called bool
-	server := httptest.NewServer(http.HandlerFunc(
-		func(w http.ResponseWriter, req *http.Request) {
-			called = true
-
-			w.Write([]byte(c.Output))
-
-			assert.Equal(t, c.ExpectedPath, req.URL.String())
-			assert.Equal(t, c.ExpectedMethod, req.Method)
-		},
-	))
-
-	client, _ := api.NewClient(server.URL)
-
-	fmt.Println(client.Do(c.Schema))
-
-	assert.Equal(t, true, called, "lfs/api: expected method %s to be called", c.ExpectedPath)
-	assert.Equal(t, c.ExpectedResponse, c.Response)
+	assert.Nil(t, resp.Err)
+	assert.Equal(t, "", resp.CommitNeeded)
+	assert.Equal(t, "some-lock-id", resp.Lock.Id)
+	assert.Equal(t, "/path/to/file", resp.Lock.Path)
+	assert.Equal(t, api.Committer{"Jane Doe", "jane@example.com"}, resp.Lock.Committer)
+	assert.Equal(t, now, resp.Lock.LockedAt)
+	assert.True(t, resp.Lock.Active())
 }
